# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UPER IDP API Documentation",
        "description": "API documentation for UPER IDP SSO System (CENTRAL)",
        "contact": {
            "email": "burhan.mafazi@universitaspertamina.ac.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/applications/{uuid}/status": {
            "put": {
                "tags": [
                    "Applications"
                ],
                "summary": "Update application status",
                "operationId": "ab975aee67c000b354f66f1ef4a2f6d8",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Application UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Application status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Application status updated successfully"
                                        },
                                        "url": {
                                            "type": "string",
                                            "example": "http://localhost:8000/api/v1/applications/567b7c8b-4b5e-4b5e-4b5e-4b5e4b5e4b5e/status"
                                        },
                                        "method": {
                                            "type": "string",
                                            "example": "PUT"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "example": "2023-06-01 10:00:00"
                                        },
                                        "total_data": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ApplicationResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/applications/{uuid}/users": {
            "get": {
                "tags": [
                    "Applications"
                ],
                "summary": "Get users assigned to an application",
                "operationId": "d34dd8656de247ac36fbbefe6868f69b",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Application UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Users retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Users retrieved successfully"
                                        },
                                        "url": {
                                            "type": "string",
                                            "example": "http://localhost:8000/api/v1/applications/567b7c8b-4b5e-4b5e-4b5e-4b5e4b5e4b5e/users"
                                        },
                                        "method": {
                                            "type": "string",
                                            "example": "GET"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "example": "2023-06-01 10:00:00"
                                        },
                                        "total_data": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UserRoleResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/me/applications": {
            "get": {
                "tags": [
                    "Applications"
                ],
                "summary": "Get my applications",
                "operationId": "be3384e1c267cbf5e627f0e498e5310a",
                "responses": {
                    "200": {
                        "description": "My applications retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "My applications retrieved successfully"
                                        },
                                        "url": {
                                            "type": "string",
                                            "example": "http://localhost:8000/api/v1/auth/me/applications"
                                        },
                                        "method": {
                                            "type": "string",
                                            "example": "GET"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "example": "2023-06-01 10:00:00"
                                        },
                                        "total_data": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ApplicationResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApplicationResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "uuid": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "code": {
                        "type": "string",
                        "example": "APP001"
                    },
                    "name": {
                        "type": "string",
                        "example": "My Application"
                    },
                    "alias": {
                        "type": "string",
                        "example": "MyApp"
                    },
                    "description": {
                        "type": "string",
                        "example": "Description of the application"
                    },
                    "base_url": {
                        "type": "string",
                        "example": "https://myapp.example.com"
                    },
                    "login_url": {
                        "type": "string",
                        "example": "https://myapp.example.com/login"
                    },
                    "platform_type": {
                        "type": "string",
                        "example": "web"
                    },
                    "visibility": {
                        "type": "string",
                        "example": "public"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "image": {
                        "type": "string",
                        "example": "https://myapp.example.com/logo.png"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with username and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}